# Cache results of computing model type