--- a/activerecord/test/cases/inheritance_test.rb
+++ b/activerecord/test/cases/inheritance_test.rb
@@ -58,21 +58,21 @@ class InheritanceTest < ActiveRecord::TestCase
   end
 
   def test_compute_type_success
-    assert_equal Author, ActiveRecord::Base.send(:compute_type, "Author")
+    assert_equal Author, Company.send(:compute_type, "Author")
   end
 
   def test_compute_type_nonexistent_constant
     e = assert_raises NameError do
-      ActiveRecord::Base.send :compute_type, "NonexistentModel"
+      Company.send :compute_type, "NonexistentModel"
     end
-    assert_equal "uninitialized constant ActiveRecord::Base::NonexistentModel", e.message
-    assert_equal "ActiveRecord::Base::NonexistentModel", e.name
+    assert_equal "uninitialized constant Company::NonexistentModel", e.message
+    assert_equal "Company::NonexistentModel", e.name
   end
 
   def test_compute_type_no_method_error
     ActiveSupport::Dependencies.stub(:safe_constantize, proc { raise NoMethodError }) do
       assert_raises NoMethodError do
-        ActiveRecord::Base.send :compute_type, "InvalidModel"
+        Company.send :compute_type, "InvalidModel"
       end
     end
   end
@@ -90,7 +90,7 @@ class InheritanceTest < ActiveRecord::TestCase
     ActiveSupport::Dependencies.stub(:safe_constantize, proc { raise e }) do
 
       exception = assert_raises NameError do
-        ActiveRecord::Base.send :compute_type, "InvalidModel"
+        Company.send :compute_type, "InvalidModel"
       end
       assert_equal error.message, exception.message
     end
@@ -99,7 +99,7 @@ class InheritanceTest < ActiveRecord::TestCase
   def test_compute_type_argument_error
     ActiveSupport::Dependencies.stub(:safe_constantize, proc { raise ArgumentError }) do
       assert_raises ArgumentError do
-        ActiveRecord::Base.send :compute_type, "InvalidModel"
+        Company.send :compute_type, "InvalidModel"
       end
     end
   end