@pixeltrix 2017-01-01
First, I think we'd need to use a thread-safe cache here and not a standard ruby hash. Second, I'm not convinced that there aren't some edge conditions with `compute_type` that could result in either the wrong constant being cached or it getting corrupted in development mode when we're reloading classes.

However, there's been a few reports of slowness around this code so I think the underlying issue needs addressing. One question I have though is, why you're using an enum column instead of a varchar? The option to not store full class names is really a legacy option - IIRC it was meant to be removed at some point.