@matthewd 2017-01-01
Yeah, it needs to be a `Concurrent::Map`.

We should be okay on reloading because it's stored on the class instance -- so when we reload, the cache-owning class will go away at the same time the referred class does. We could be extra-paranoid and just cache the fully qualified name string instead of the resolved class, but I don't think that's necessary.