@Le6ow5k1 2017-01-01
Thanks for pointing out issues.

> First, I think we'd need to use a thread-safe cache here and not a standard ruby hash

`Thread.current[]` can be used, but each class needs to have it's own `type_candidates`. Something like this might work:

```ruby
Thread.current[name] ||= { type_candidates: {} }
```

> Second, I'm not convinced that there aren't some edge conditions with compute_type that could result in either the wrong constant being cached or it getting corrupted in development mode when we're reloading classes.

I think the option here is to store not the constants but strings representing them. So if cached type is present we still need to `constantize` it and if this fails (in case of reloading classes) we continue by calculating new type.

```ruby
candidates_cache = Thread.current[name][:type_candidates]
cached_candidate = candidates_cache[type_name]
return type_constant \
  if type_constant = ActiveSupport::Dependencies.safe_constantize(cached_candidate)
...
```

> One question I have though is, why you're using an enum column instead of a varchar?

Because there is no reason to store duplicate type names in every row. With enum type only finite number of values are stored on disk. It can save a lot of space if you have millions of rows in table. Also enum gives us type safety so only specific set of values is allowed in enum column, no need for additional constraint.