@pixeltrix 2017-01-02
> We should be okay on reloading because it's stored on the class instance -- so when we reload, the cache-owning class will go away at the same time the referred class does. We could be extra-paranoid and just cache the fully qualified name string instead of the resolved class, but I don't think that's necessary.

Not sure given the changes in the area of code reloading in Rails 5.0, but this will only be true if all classes are unloaded and reloaded. Since `compute_type` is used to determine the correct class for an association as well as in STI then if only the association target class is reloaded then you could have a reference to a stale class definition. Since `ActiveSupport::Dependencies.constantize` has it's own class caching mechanism I think it'd be safer to just cache the name and not the actual class.