@pixeltrix 2017-01-02
> It can save a lot of space if you have millions of rows in table

@Le6ow5k1 given that default enum name limit is 63 characters and you're using `store_full_sti_class == false` I'm assuming you have some very long class names that make that optimisation worthwhile 😄