@Le6ow5k1 2017-01-02
Let's say you have 50 mln rows in the table and class name length is 70 chars. If we assume that 1 char = 1 byte then `(50_000_000 * 70) / (1024 * 1024 * 1024) = 3,3 GB`. 3.3 GB if I'm counting correctly, for no good reason. And disk space is just one downside of varchar. IMHO overall enum is more appropriate type in case of STI.