@pixeltrix 2017-01-03
@Le6ow5k1 can you squash your commits and add a CHANGELOG entry, thanks. Once you've done that it's ready to merge.